+++ src/render/common/types.h
@@ -17,9 +17,9 @@ struct Color {
 };
 
 struct Vector3d {
-  double x;
-  double y;
-  double z;
+  float x;
+  float y;
+  float z;
 };
 
 struct Ray {
@@ -27,11 +27,11 @@ struct Ray {
   Vector3d direction;
 };
 
-MAYBE_DEVICE inline Vector3d operator*(const Vector3d& vec, double scalar) {
+MAYBE_DEVICE inline Vector3d operator*(const Vector3d& vec, float scalar) {
   return {vec.x * scalar, vec.y * scalar, vec.z * scalar};
 }
 
-MAYBE_DEVICE inline Vector3d operator/(const Vector3d& vec, double scalar) {
+MAYBE_DEVICE inline Vector3d operator/(const Vector3d& vec, float scalar) {
   return {vec.x / scalar, vec.y / scalar, vec.z / scalar};
 }
 
@@ -47,7 +47,7 @@ MAYBE_DEVICE inline Vector3d operator*(const Vector3d& a, const Vector3d& b) {
   return {a.x * b.x, a.y * b.y, a.z * b.z};
 }
 
-MAYBE_DEVICE inline double Dot(const Vector3d& a, const Vector3d& b) {
+MAYBE_DEVICE inline float Dot(const Vector3d& a, const Vector3d& b) {
   return a.x * b.x + a.y * b.y + a.z * b.z;
 }
 
@@ -55,7 +55,7 @@ MAYBE_DEVICE inline Vector3d Cross(const Vector3d& a, const Vector3d& b) {
   return {a.y * b.z - a.z * b.y, a.z * b.x - a.x * b.z, a.x * b.y - a.y * b.x};
 }
 
-MAYBE_DEVICE inline double Length(const Vector3d& vec) {
+MAYBE_DEVICE inline float Length(const Vector3d& vec) {
   return sqrt(Dot(vec, vec));
 }
 
@@ -71,6 +71,6 @@ MAYBE_DEVICE inline Vector3d Fract(const Vector3d& vec) {
   return {vec.x - floor(vec.x), vec.y - floor(vec.y), vec.z - floor(vec.z)};
 }
 
-MAYBE_DEVICE inline double MaxComponent(const Vector3d& vec) {
+MAYBE_DEVICE inline float MaxComponent(const Vector3d& vec) {
   return fmax(vec.x, fmax(vec.y, vec.z));
 }
